Refactor studio client
